Removing cert_chain and signing_key from gemspec
If you use gem signing, those lines should not be committed to the
gemspec in the repo for use in Bundler. They should be used when
generating the spec for bundling, but this task is only used for
recording the gemspec for use with Bundler.

This change stops writing those lines.

An additional task has been added to generate the full gemspec,
including lines related to gem signing.